Move the listener service to its own module
