+++ pass-authz-integration/pom.xml
@@ -129,7 +129,7 @@
             <pass.fedora.baseurl>http://${FCREPO_HOST}:${FCREPO_PORT}/fcrepo/rest/</pass.fedora.baseurl>
             <pass.elasticsearch.url>http://${ES_HOST}:${ES_PORT}</pass.elasticsearch.url>
             <jms.brokerUrl>failover:(tcp://localhost:${JMS_PORT})?trackMessages=true</jms.brokerUrl>
-            <authz.listener.jar>${org.dataconservancy.pass:pass-authz-tools:jar:listener-exe}</authz.listener.jar>
+            <authz.listener.jar>${org.dataconservancy.pass:pass-authz-listener:jar:exe}</authz.listener.jar>
             <pass.authz.queue>Consumer.authz.VirtualTopic.pass.authz</pass.authz.queue>
             <jms.username>${activemq.user}</jms.username>
             <jms.password>${activemq.password}</jms.password>
@@ -358,10 +358,10 @@
     
     <dependency>
       <groupId>org.dataconservancy.pass</groupId>
-      <artifactId>pass-authz-tools</artifactId>
+      <artifactId>pass-authz-listener</artifactId>
       <version>${project.version}</version>
       <scope>test</scope>
-      <classifier>listener-exe</classifier>
+      <classifier>exe</classifier>
     </dependency>
 
     <dependency>
