+++ pass-authz-listener/src/main/java/org/dataconservancy/pass/authz/listener/FedoraAction.java
@@ -1,4 +1,4 @@
-package org.dataconservancy.pass.authz.tools;
+package org.dataconservancy.pass.authz.listener;
 
 public enum FedoraAction {
     CREATED, MODIFIED, DELETED;
