+++ pom.xml
@@ -12,6 +12,7 @@
     <module>pass-authz-filters</module>
     <module>pass-authz-acl</module>
     <module>pass-authz-tools</module>
+    <module>pass-authz-listener</module>
   </modules>
 
   <properties>
